Remove envtest-existing from upgradecheck
`envtest-existing` tests have been flaky and we are moving
towards KUTTL tests for e2e PostgresCluster behavior; several
tests in the `upgradecheck` package were originally written as
`envtest-existing` but are not really suitable as KUTTL tests,
so this PR changes them from `envtest-existing` to `envtest`

Issue [sc-14243]
